requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=4.9.0=pyh29332c3_0
appnope=0.1.4=pyhd8ed1ab_1
argon2-cffi=23.1.0=pyhd8ed1ab_1
argon2-cffi-bindings=21.2.0=py312h024a12e_5
arrow=1.3.0=pyhd8ed1ab_1
arviz=0.21.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-lru=2.0.5=pyh29332c3_0
attrs=25.3.0=pyh71513ae_0
babel=2.17.0=pyhd8ed1ab_0
beautifulsoup4=4.13.3=pyha770c72_0
bleach=6.2.0=pyh29332c3_4
bleach-with-css=6.2.0=h82add2a_4
bottleneck=1.4.2=py312h147345f_0
brotli=1.1.0=hd74edd7_2
brotli-bin=1.1.0=hd74edd7_2
brotli-python=1.1.0=py312hde4cb15_2
bspline=0.1.1=pypi_0
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.1.31=hf0a4a13_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.5.2=pypi_0
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py312h0fad829_0
charset-normalizer=3.4.1=pyhd8ed1ab_0
cloudpickle=3.1.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
cons=0.4.6=pypi_0
contourpy=1.3.1=py312hb23fbb9_0
cycler=0.12.1=pyhd8ed1ab_1
debugpy=1.8.13=py312hd8f9ff3_0
decorator=5.2.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
etuples=0.3.9=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
filelock=3.18.0=pypi_0
fonttools=4.56.0=py312h998013c_0
fqdn=1.5.1=pyhd8ed1ab_1
freetype=2.13.3=h1d14073_0
fsspec=2025.3.0=pypi_0
h11=0.14.0=pyhd8ed1ab_1
h2=4.2.0=pyhd8ed1ab_0
h5netcdf=1.6.1=pypi_0
h5py=3.13.0=pypi_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.7=pyh29332c3_1
httpx=0.28.1=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.6.1=pyha770c72_0
importlib_resources=6.5.2=pyhd8ed1ab_0
ipykernel=6.29.5=pyh57ce528_0
ipython=9.0.2=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
isoduration=20.11.0=pyhd8ed1ab_1
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_1
joypy=0.2.4=pyhd8ed1ab_1
json5=0.10.0=pyhd8ed1ab_1
jsonpointer=3.0.0=py312h81bd7bf_1
jsonschema=4.23.0=pyhd8ed1ab_1
jsonschema-specifications=2024.10.1=pyhd8ed1ab_1
jsonschema-with-format-nongpl=4.23.0=hd8ed1ab_1
jupyter-lsp=2.2.5=pyhd8ed1ab_1
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
jupyter_events=0.12.0=pyh29332c3_0
jupyter_server=2.15.0=pyhd8ed1ab_0
jupyter_server_terminals=0.5.3=pyhd8ed1ab_1
jupyterlab=4.3.6=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_2
jupyterlab_server=2.27.3=pyhd8ed1ab_1
kiwisolver=1.4.8=py312h2c4a281_0
krb5=1.21.3=h237132a_0
lcms2=2.17=h7eeda09_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=31_h10e41b3_openblas
libbrotlicommon=1.1.0=hd74edd7_2
libbrotlidec=1.1.0=hd74edd7_2
libbrotlienc=1.1.0=hd74edd7_2
libcblas=3.9.0=31_hb3479ef_openblas
libcxx=20.1.1=ha82da77_0
libdeflate=1.23=hec38601_0
libedit=3.1.20250104=pl5321hafb1f1b_0
libexpat=2.6.4=h286801f_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=14_2_0_h6c33f7e_1
libgfortran5=14.2.0=h6c33f7e_1
libiconv=1.18=hfe07756_1
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=31_hc9a63f6_openblas
liblzma=5.6.4=h39f12f2_0
libopenblas=0.3.29=openmp_hf332438_0
libpng=1.6.47=h3783ad8_0
libsodium=1.0.20=h99b78c6_0
libsqlite=3.49.1=h3f77e49_2
libtiff=4.7.0=h551f018_3
libwebp-base=1.5.0=h2471fea_0
libxcb=1.17.0=hdb1d25a_0
libxml2=2.13.6=hce475f1_0
libxslt=1.1.39=h223e5b9_0
libzlib=1.3.1=h8359307_2
littleutils=0.2.4=pyhd8ed1ab_1
llvm-openmp=20.1.1=hdb05f8b_1
llvmlite=0.43.0=pypi_0
logical-unification=0.4.6=pypi_0
lxml=5.3.1=py312h9535dd2_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py312h998013c_1
matplotlib-base=3.10.1=py312hdbc7e53_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
minikanren=1.0.3=pypi_0
mistune=3.1.3=pyh29332c3_0
mpmath=1.3.0=pyhd8ed1ab_1
multipledispatch=1.0.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
nbclient=0.10.2=pyhd8ed1ab_0
nbconvert-core=7.16.6=pyh29332c3_0
nbformat=5.10.4=pyhd8ed1ab_1
ncurses=6.5=h5e97a16_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
nibabel=5.3.2=pyha770c72_1
nilearn=0.11.1=pyhd8ed1ab_0
notebook-shim=0.2.4=pyhd8ed1ab_1
numba=0.60.0=pypi_0
numexpr=2.10.2=py312hbbbb429_0
numpy=2.0.2=pypi_0
nutpie=0.13.4=pypi_0
openjpeg=2.5.3=h8a3d83b_0
openssl=3.4.1=h81ee809_0
outdated=0.2.2=pyhd8ed1ab_2
overrides=7.7.0=pyhd8ed1ab_1
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py312hcf29cfe_0
pandas-flavor=0.6.0=pyhd8ed1ab_3
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_1
patsy=1.0.1=pyhd8ed1ab_1
pcntoolkit=0.33.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=py312h50aef2c_0
pingouin=0.5.5=pyhd8ed1ab_1
pip=25.0.1=pyh8b19718_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_2
platformdirs=4.3.7=pyh29332c3_0
prometheus_client=0.21.1=pyhd8ed1ab_0
prompt-toolkit=3.0.50=pyha770c72_0
psutil=7.0.0=py312hea69d52_0
pthread-stubs=0.4=hd74edd7_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=19.0.1=pypi_0
pycparser=2.22=pyh29332c3_1
pygments=2.19.1=pyhd8ed1ab_0
pymc=5.21.1=pypi_0
pyobjc-core=11.0=py312hb9d441b_0
pyobjc-framework-cocoa=11.0=py312hb9d441b_0
pyparsing=3.2.2=pyhd8ed1ab_0
pysocks=1.7.1=pyha55dd90_7
pytensor=2.28.3=pypi_0
python=3.12.9=hc22306f_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-fastjsonschema=2.21.1=pyhd8ed1ab_0
python-json-logger=2.0.7=pyhd8ed1ab_0
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.12=5_cp312
pytz=2025.1=pyhd8ed1ab_0
pyyaml=6.0.2=py312h998013c_2
pyzmq=26.3.0=py312hf4875e0_0
qhull=2020.2=h420ef59_5
readline=8.2=h1d1bf99_2
referencing=0.36.2=pyh29332c3_0
requests=2.32.3=pyhd8ed1ab_1
rfc3339-validator=0.1.4=pyhd8ed1ab_1
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rich=13.9.4=pypi_0
rpds-py=0.23.1=py312hd60eec9_0
scikit-learn=1.6.1=py312h39203ce_0
scikit-posthocs=0.11.4=pypi_0
scipy=1.15.2=py312h99a188d_0
seaborn=0.13.2=hd8ed1ab_3
seaborn-base=0.13.2=pyhd8ed1ab_3
send2trash=1.8.3=pyh31c8845_1
setuptools=75.8.2=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pyhd8ed1ab_1
soupsieve=2.5=pyhd8ed1ab_1
stack_data=0.6.3=pyhd8ed1ab_1
statannot=0.2.3=pypi_0
statannotations=0.7.2=pypi_0
statsmodels=0.14.4=py312h755e627_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pyhd8ed1ab_2
terminado=0.18.1=pyh31c8845_0
threadpoolctl=3.6.0=pyhecae5ae_0
tinycss2=1.4.0=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tomli=2.2.1=pyhd8ed1ab_1
toolz=1.0.0=pypi_0
torch=2.6.0=pypi_0
tornado=6.4.2=py312hea69d52_0
tqdm=4.67.1=py312h989b03a_0
traitlets=5.14.3=pyhd8ed1ab_1
types-python-dateutil=2.9.0.20241206=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
typing_utils=0.1.0=pyhd8ed1ab_1
tzdata=2025b=h78e105d_0
unicodedata2=16.0.0=py312hea69d52_0
uri-template=1.3.0=pyhd8ed1ab_1
urllib3=2.3.0=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_1
webcolors=24.11.1=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_3
websocket-client=1.8.0=pyhd8ed1ab_1
wheel=0.45.1=pyhd8ed1ab_1
xarray=2025.3.0=pyhd8ed1ab_0
xarray-einstats=0.8.0=pypi_0
xorg-libxau=1.0.12=h5505292_0
xorg-libxdmcp=1.1.5=hd74edd7_0
yaml=0.2.5=h3422bc3_2
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_1
zstandard=0.23.0=py312hea69d52_1
zstd=1.5.7=h6491c7d_2
